requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.3.0
autocommand==2.2.2
backports.tarfile==1.2.0
blinker==1.9.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
cheroot==10.0.1
CherryPy==18.10.0
click==8.1.8
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dash==3.0.0
decorator==5.2.1
distlib==0.3.9
docker-pycreds==0.4.0
einops==0.8.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
flash_attn==2.7.4.post1
Flask==3.0.3
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
git-filter-repo==2.38.0
gitdb==4.0.12
GitPython==3.1.44
h5py==3.13.0
huggingface-hub==0.29.3
hydra-core==1.3.2
identify==2.6.9
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
iopath==0.1.10
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lightning==2.5.0.post0
lightning-utilities==0.14.1
lmdb==1.6.2
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
narwhals==1.31.0
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.9.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open3d==0.19.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==6.0.1
plyfile==1.1
portalocker==3.1.1
portend==3.2.0
pre_commit==4.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyceres==2.4
pycolmap==3.11.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pyquaternion==0.9.9
pyte==0.8.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==13.9.4
roma==1.5.2.1
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.23.1
setproctitle==1.3.5
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
stringcase==1.2.0
stringzilla==3.12.3
sympy==1.13.1
tempora==5.8.0
tensorboardX==2.6.2.2
thefuck==3.32
threadpoolctl==3.6.0
timm==1.0.15
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.6.3
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transforms3d==0.4.2
trimesh==4.6.4
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.3
visdom==0.2.4
vulture==2.14
wandb==0.19.8
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wis3d==1.0.1
yarl==1.18.3
zc.lockfile==3.0.post1
zipp==3.21.0
